requirements.txt
flask==3.1.2
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.7.1
flask_mail==0.10.0
pika==1.3.2
requests==2.32.5
psycopg2-binary==2.9.10
python-dotenv==1.1.1
flask-dance==7.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
flask-login==0.6.3
weasyprint==66.0
pytz==2025.2
gunicorn==23.0.0